Revert AdviceCard changes
