--- packages/js/product-editor/src/components/advice-card/index.ts
@@ -1,2 +1,0 @@
-export * from './advice-card';
-export * from './types';
