+++ packages/js/product-editor/src/components/advice-card/index.tsx
@@ -1,17 +1,25 @@
 /**
  * External dependencies
  */
-import { Button } from '@wordpress/components';
 import { createElement } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
+import {
+	Button,
+	Card,
+	CardBody,
+	CardFooter,
+	CardHeader,
+} from '@wordpress/components';
 import { close } from '@wordpress/icons';
 import { useInstanceId } from '@wordpress/compose';
 import { useUserPreferences } from '@woocommerce/data';
 
-/**
- * Internal dependencies
- */
-import { AdviceCardProps } from './types';
+export interface AdviceCardProps {
+	tip?: string;
+	isDismissible?: boolean;
+	onDismiss?: () => void;
+	image?: React.ReactNode;
+}
 
 export const AdviceCard: React.FC< AdviceCardProps > = ( {
 	tip,
@@ -48,27 +56,19 @@ export const AdviceCard: React.FC< AdviceCardProps > = ( {
 	}
 
 	return (
-		<div className="woocommerce-advice-card">
-			<div className="woocommerce-advice-card__body">
-				{ isDismissible && (
+		<Card className="woocommerce-advice-card">
+			{ isDismissible && (
+				<CardHeader>
 					<Button
 						className="woocommerce-advice-card__dismiss-button"
 						onClick={ onDismiss }
 						icon={ close }
 						label={ __( 'Dismiss', 'woocommerce' ) }
 					/>
-				) }
-
-				{ image && (
-					<div className="woocommerce-advice-card__image">
-						{ image }
-					</div>
-				) }
-
-				{ Boolean( tip?.length ) && (
-					<p className="woocommerce-advice-card__tip">{ tip }</p>
-				) }
-			</div>
-		</div>
+				</CardHeader>
+			) }
+			<CardBody>{ image }</CardBody>
+			{ tip && tip.length > 0 && <CardFooter>{ tip }</CardFooter> }
+		</Card>
 	);
 };
