--- packages/js/product-editor/src/components/advice-card/types.ts
@@ -1,6 +1,0 @@
-export interface AdviceCardProps {
-	image?: React.ReactNode;
-	tip?: string;
-	isDismissible?: boolean;
-	onDismiss?: () => void;
-}
